Antonio13072020

+++ clientSviluppo/src/main/java/it/csi/demetra/demetraws/zoo/model/Dmt_t_output_controlli.java
@@ -49,14 +49,14 @@ public class Dmt_t_output_controlli implements Serializable {
 	/**
 	 * Numero di capi richiesti a premio dal richiedente.
 	 */
-	private Integer capiRichiesti;
+	private BigDecimal capiRichiesti;
 
 	@Column(name = "capi_ammissibili")
 	/**
 	 * capi ammessi a premio
 	 */
 	private BigDecimal capiAmmissibili;
-
+		
 	/**
 	 * Metodo getter che ritorna l'identificativo univoco associato all'esecuzione.
 	 * @return idSessione identificativo univoco associato all'esecuzione.
@@ -125,7 +125,7 @@ public class Dmt_t_output_controlli implements Serializable {
 	 * Metodo getter che ritorna il numero di capi richiesti dal richiedente.
 	 * @return capiRichiesti numero di capi richiesti dal richiedente.
 	 */
-	public Integer getCapiRichiesti() {
+	public BigDecimal getCapiRichiesti() {
 		return capiRichiesti;
 	}
 
@@ -133,7 +133,7 @@ public class Dmt_t_output_controlli implements Serializable {
 	 * Metodo setter che imposta il numero di capi richiesti dal richiedente.
 	 * @param capiRichiesti numero di capi richiesti dal richiedente.
 	 */
-	public void setCapiRichiesti(Integer capiRichiesti) {
+	public void setCapiRichiesti(BigDecimal capiRichiesti) {
 		this.capiRichiesti = capiRichiesti;
 	}
